pyproject.toml
[tool.poetry]
name = "whisper-local"
version = "0.1.0"
description = "Local Whisper instance with API server"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
openai-whisper = "^20231117"
python-multipart = "^0.0.6"
pydantic = "^2.5.2"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
